# internal/api/openapi.yaml
openapi: "3.0.3"
info:
  title: ClawBench API
  description: |
    ClawBench 的完整 HTTP API。所有 `/api/` 路由默认经 `middleware.Auth`
    鉴权（会话 Cookie，或本机 AI 子进程的短时 `aiToken`），少数端点刻意免鉴权
    （见各 operation 的 `security: []`）。
    聊天流式内容通过 WebSocket 推送（见 `/api/ai/events/ws`），不在本文件建模。

    约定：
    - 统一错误体见 `components/schemas/ErrorResponse`。
    - 项目范围通过 `clawbench_project` Cookie 或 `project_path` 查询参数传递。
    - 时间戳为 epoch 毫秒（除显式说明外）。
  version: "1.0.0"
servers:
  - url: /
    description: 同源（文件管理器内的 Swagger UI 直接可用）
...
security:
  - cookieAuth: []
  - aiToken: []
paths:
...
  /api/usage/stats:
    get:
      tags: [System]
      summary: 用量统计
      description: |
        按维度聚合 `chat_metadata` 的 token / 成本统计。

        **项目范围**由 `scope` 决定：
        - `scope=project`（默认）：限制在项目 Cookie 指定的单个项目内，Cookie 必填。
        - `scope=all`：聚合实例内**全部项目**。任何已鉴权调用方均可用（浏览器面板
          的「所有项目」开关、`/cb-usage` 的本机 AI 令牌）；未鉴权者被拒绝。该项目
          Cookie 被忽略，跨项目统计通常配合 `dims=project`。

        `dims` 合法值含 `project`，但只在 `scope=all` 下有区分度——单项目范围内所有
        行项目相同。

        **响应结构**：
        - `totals`：所有行的合计，字段 `input` / `output` / `total` / `cacheHit` /
          `cacheMiss` / `credit` / `costUsd` / `messageCnt`。
        - `rows[]`：每个 dim 组合一行。`key` 是对象，键即请求的 dims
          （如 `{"model":"...","backend":"..."}`），行上带有与 `totals` 相同的字段。
          行的顺序由 `sort` / `order` 决定，`limit` 截断。
        - `trend[]`：仅 `trend=1` 时返回，在 `rows[]` 的基础上多一个 `day`
          （`YYYY-MM-DD`），用于按天走势图。

        注意 `chat_metadata.created_at` 记录的是元数据**最终写入**时刻
        （约等于助手回复完成），不是用户提问时刻。
      operationId: usageStats
      parameters:
        - name: scope
          in: query
          description: |
            统计范围。默认 `project`（限制在项目 Cookie 指定的项目内）。
            `all` 聚合实例内全部项目，任何已鉴权调用方均可用；该模式下项目
            Cookie 被忽略。
          schema:
            type: string
            enum: [project, all]
        - name: start
          in: query
          required: true
          description: 起始时间（RFC3339）。
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          required: true
          description: 结束时间（RFC3339）。
          schema:
            type: string
            format: date-time
        - name: dims
          in: query
          required: true
          description: |
            聚合维度。**重复该参数**以多维分组（1..4 个，不可重复）——
            例如 `dims=model&dims=backend`。逗号分隔会被拒绝。
            合法值为 `model` / `backend` / `agent` / `project`；`project`
            只在 `scope=all` 下有区分度。
          schema:
            type: array
            items:
              type: string
              enum: [model, backend, agent, project]
        - name: metrics
          in: query
          required: true
          description: |
            指标。**重复该参数**（至少 1 个）——例如 `metrics=input&metrics=total`。
            逗号分隔会被拒绝。
          schema:
            type: array
            items:
              type: string
              enum: [input, output, total, cacheHit, credit, cost]
        - name: sort
          in: query
          description: 排序用的指标 id，默认 `total`。
          schema:
            type: string
        - name: order
          in: query
          description: 排序方向，默认 `desc`。
          schema:
            type: string
            enum: [asc, desc]
        - name: trend
          in: query
          description: 传 `1` 时额外返回按天分组的 `trend[]`。
          schema:
            type: string
        - name: top
          in: query
          description: trend 模式下保留的 dim 组合数，默认 10。
          schema:
            type: integer
        - name: limit
          in: query
          description: 最大行数，默认 50，上限 200。
          schema:
            type: integer
      responses:
        "200":
          description: 聚合结果
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: clawbench_session
      description: 登录后下发的会话 Cookie；CLI 使用 `<DataDir>/cookie-token`。
    aiToken:
      type: apiKey
      in: header
      name: X-ClawBench-AI-Token
      description: |
        本机 AI 子进程使用的短时令牌，由内置斜杠命令（`/cb-task` 等）在注入提示词时签发，
        有效期 30 分钟，以 `CookieToken` 做 HMAC 签名。仅接受来自回环地址的请求：
        地址与签名须同时成立。地址单独不再放行——FRP 隧道的 frpc 亦从 `127.0.0.1` 回拨。
...